Remove BFF and create auth module
