--- charts/dbildungs-iam/templates/dbildungs-iam-service-bff.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: dbiam-bff
-  labels:
-    app.kubernetes.io/name: {{.Chart.Name}}
-    app.kubernetes.io/version: {{.Chart.Version}}
-spec:
-  selector:
-    layer: dbildungs-iam-bff
-  ports:
-    - protocol: TCP
-      name: web
-      port: {{.Values.dbildungsIamExternalPort}}
-      targetPort: web
-  type: ClusterIP
