+++ compose.yaml
@@ -44,26 +44,6 @@ services:
       - source: 'config'
         target: '/app/config'
         type: bind
-  bff:
-    profiles:
-      - full-backend
-    image: ghcr.io/dbildungsplattform/dbildungs-iam-server
-    build:
-      context: .
-    depends_on:
-      - backend
-      - redis
-      - keycloak
-    restart: on-failure
-    environment:
-      NODE_ENV: dev
-      DEPLOY_STAGE: dev
-    command: ['node', 'dist/src/backend-for-frontend/main.js']
-    network_mode: host
-    volumes:
-      - source: 'config'
-        target: '/app/config'
-        type: bind
   db-init:
     profiles:
       - db-init
