+++ nest-cli.json
@@ -19,15 +19,6 @@
                 "tsConfigPath": "tsconfig.build.json"
             }
         },
-        "backend-for-frontend": {
-            "type": "application",
-            "root": "src",
-            "entryFile": "src/backend-for-frontend/main",
-            "sourceRoot": "src",
-            "compilerOptions": {
-                "tsConfigPath": "tsconfig.build.json"
-            }
-        },
         "console": {
             "type": "application",
             "root": "src",
