+++ package.json
@@ -15,9 +15,6 @@
         "start": "nest start server ",
         "start:debug": "nest start server --debug --watch",
         "start:prod": "node dist/src/server/main.js",
-        "start:bff": "nest start backend-for-frontend",
-        "start:bff:debug": "nest start backend-for-frontend --debug",
-        "start:bff:prod": "node dist/src/backend-for-frontend/main.js",
         "test": "jest",
         "test:ci": "jest --coverage --maxWorkers=100%",
         "test:cov": "jest --coverage --maxWorkers=50%",
