--- src/backend-for-frontend/backend-for-frontend.module.ts
@@ -1,50 +1,0 @@
-import { classes } from '@automapper/classes';
-import { CamelCaseNamingConvention } from '@automapper/core';
-import { AutomapperModule } from '@automapper/nestjs';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
-import { defineConfig } from '@mikro-orm/postgresql';
-import { Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-
-import { FrontendApiModule } from '../modules/frontend/frontend-api.module.js';
-import { DbConfig, loadConfigFiles, ServerConfig } from '../shared/config/index.js';
-import { mappingErrorHandler } from '../shared/error/mapping.error.js';
-import { HealthModule } from '../modules/health/health.module.js';
-import { KeycloakConfigModule } from '../modules/keycloak-administration/keycloak-config.module.js';
-
-@Module({
-    imports: [
-        ConfigModule.forRoot({
-            isGlobal: true,
-            load: [loadConfigFiles],
-        }),
-        AutomapperModule.forRoot({
-            strategyInitializer: classes(),
-            namingConventions: new CamelCaseNamingConvention(),
-            errorHandler: mappingErrorHandler,
-        }),
-        MikroOrmModule.forRootAsync({
-            useFactory: (config: ConfigService<ServerConfig, true>) => {
-                const dbConfig: DbConfig = config.getOrThrow<DbConfig>('DB');
-                return defineConfig({
-                    clientUrl: dbConfig.CLIENT_URL,
-                    dbName: dbConfig.DB_NAME,
-                    entities: ['./dist/**/*.entity.js'],
-                    entitiesTs: ['./src/**/*.entity.ts'],
-                    // Needed for HealthCheck
-                    type: 'postgresql',
-                    driverOptions: {
-                        connection: {
-                            ssl: dbConfig.USE_SSL,
-                        },
-                    },
-                });
-            },
-            inject: [ConfigService],
-        }),
-        FrontendApiModule,
-        HealthModule,
-        KeycloakConfigModule,
-    ],
-})
-export class BackendForFrontendModule {}
