+++ src/modules/authentication/api/frontend.controller.spec.ts
@@ -13,10 +13,10 @@ import {
 
 import { ConfigTestModule } from '../../../../test/utils/config-test.module.js';
 import { FrontendConfig } from '../../../shared/config/frontend.config.js';
-import { OIDC_CLIENT } from '../auth/oidc-client.service.js';
-import { User } from '../auth/user.decorator.js';
+import { OIDC_CLIENT } from '../services/oidc-client.service.js';
+import { User } from '../services/user.decorator.js';
 import { ProviderService } from '../outbound/provider.service.js';
-import { FrontendController } from './frontend.controller.js';
+import { AuthenticationController } from './frontend.controller.js';
 import { GetServiceProviderInfoDo } from '../../rolle/domain/get-service-provider-info.do.js';
 import { PersonService } from '../outbound/person.service.js';
 import { PersonendatensatzResponse } from '../../person/api/personendatensatz.response.js';
@@ -36,8 +36,6 @@ import { OrganisationsTyp } from '../../organisation/domain/organisation.enums.j
 import { OrganisationResponse } from '../../organisation/api/organisation.response.js';
 import { firstValueFrom, of } from 'rxjs';
 import { FindOrganisationQueryParams } from '../../organisation/api/find-organisation-query.param.js';
-import { RolleResponse } from '../../rolle/api/rolle.response.js';
-import { RolleService } from '../outbound/rolle.service.js';
 
 function getPersonenDatensatzResponse(): PersonendatensatzResponse {
     const mockBirthParams: PersonBirthParams = {
@@ -76,33 +74,30 @@ function getPersonenDatensatzResponse(): PersonendatensatzResponse {
 }
 describe('FrontendController', () => {
     let module: TestingModule;
-    let frontendController: FrontendController;
+    let frontendController: AuthenticationController;
     let oidcClient: DeepMocked<Client>;
     let frontendConfig: FrontendConfig;
     let providerService: DeepMocked<ProviderService>;
     let personService: DeepMocked<PersonService>;
     let organisationService: DeepMocked<OrganisationService>;
-    let rolleService: DeepMocked<RolleService>;
 
     beforeAll(async () => {
         module = await Test.createTestingModule({
             imports: [ConfigTestModule],
             providers: [
-                FrontendController,
+                AuthenticationController,
                 { provide: ProviderService, useValue: createMock<ProviderService>() },
                 { provide: PersonService, useValue: createMock<PersonService>() },
                 { provide: OrganisationService, useValue: createMock<OrganisationService>() },
-                { provide: RolleService, useValue: createMock<RolleService>() },
                 { provide: OIDC_CLIENT, useValue: createMock<Client>() },
             ],
         }).compile();
 
-        frontendController = module.get(FrontendController);
+        frontendController = module.get(AuthenticationController);
         oidcClient = module.get(OIDC_CLIENT);
         providerService = module.get(ProviderService);
         personService = module.get(PersonService);
         organisationService = module.get(OrganisationService);
-        rolleService = module.get(RolleService);
         frontendConfig = module.get(ConfigService).getOrThrow<FrontendConfig>('FRONTEND');
     });
 
@@ -319,6 +314,7 @@ describe('FrontendController', () => {
                     vorname: faker.person.firstName(),
                 };
                 const createPersonBodyParams: CreatePersonBodyParams = {
+                    username: faker.string.alpha(),
                     mandant: faker.string.alpha(),
                     name: personNameParams,
                 };
@@ -476,23 +472,6 @@ describe('FrontendController', () => {
             );
 
             expect(result).toEqual(pagedResponse);
-        });
-    });
-
-    describe('createRolle', () => {
-        it('should return created rolle', async () => {
-            const rolle: RolleResponse = {
-                id: faker.string.uuid(),
-                name: faker.hacker.noun(),
-                administeredBySchulstrukturknoten: faker.string.uuid(),
-                createdAt: faker.date.recent(),
-                updatedAt: faker.date.recent(),
-            };
-            rolleService.createRolle.mockResolvedValueOnce(rolle);
-
-            const result: RolleResponse = await frontendController.createRolle(createMock(), createMock<User>());
-
-            expect(result).toEqual(rolle);
         });
     });
 });
