+++ src/modules/authentication/declarations/request.d.ts
@@ -1,0 +1,11 @@
+import { PassportUser, User } from '../types/user.js';
+
+declare module 'express' {
+    interface Request {
+        user?: User;
+        accessTokenJWT?: string;
+        passportUser?: PassportUser;
+    }
+}
+
+export {};
