+++ src/modules/authentication/declarations/session.d.ts
