+++ src/modules/authentication/passport/oidc.strategy.spec.ts
@@ -5,7 +5,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { AuthorizationParameters, Client, Issuer, TokenSet, UserinfoResponse } from 'openid-client';
 
 import { ConfigTestModule } from '../../../../test/utils/index.js';
-import { OIDC_CLIENT } from './oidc-client.service.js';
+import { OIDC_CLIENT } from '../services/oidc-client.service.js';
 import { OpenIdConnectStrategy } from './oidc.strategy.js';
 
 describe('OpenIdConnectStrategy', () => {
