+++ src/modules/authentication/passport/oidc.strategy.ts
@@ -4,7 +4,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { AuthorizationParameters, Client, Strategy, StrategyOptions, TokenSet, UserinfoResponse } from 'openid-client';
 
 import { FrontendConfig, ServerConfig } from '../../../shared/config/index.js';
-import { OIDC_CLIENT } from './oidc-client.service.js';
+import { OIDC_CLIENT } from '../services/oidc-client.service.js';
 
 @Injectable()
 export class OpenIdConnectStrategy extends PassportStrategy(Strategy, 'oidc') {
