+++ src/modules/authentication/passport/session.serializer.spec.ts
