+++ src/modules/authentication/passport/session.serializer.ts
