--- src/modules/frontend/auth/index.ts
@@ -1,6 +1,0 @@
-export * from './authenticated.guard.js';
-export * from './login.guard.js';
-export * from './oidc-client.service.js';
-export * from './oidc.strategy.js';
-export * from './session.serializer.js';
-export * from './user.decorator.js';
