+++ src/server/main.ts
@@ -1,16 +1,21 @@
 /* eslint-disable no-console */
 import { NestFactory } from '@nestjs/core';
-import { INestApplication } from '@nestjs/common';
+import { NestExpressApplication } from '@nestjs/platform-express';
 import { ConfigService } from '@nestjs/config';
 import { DocumentBuilder, OpenAPIObject, SwaggerModule } from '@nestjs/swagger';
+import RedisStore from 'connect-redis';
+import session from 'express-session';
+import passport from 'passport';
+import { RedisClientType, createClient } from 'redis';
 import { NestLogger } from '../core/logging/nest-logger.js';
-import { HostConfig, ServerConfig } from '../shared/config/index.js';
+import { HostConfig, RedisConfig, ServerConfig } from '../shared/config/index.js';
 import { GlobalValidationPipe } from '../shared/validation/index.js';
 import { ServerModule } from './server.module.js';
 import { GlobalPagingHeadersInterceptor } from '../shared/paging/index.js';
+import { sessionAccessTokenMiddleware } from '../modules/authentication/services/session-access-token.middleware.js';
 
 async function bootstrap(): Promise<void> {
-    const app: INestApplication = await NestFactory.create(ServerModule);
+    const app: NestExpressApplication = await NestFactory.create<NestExpressApplication>(ServerModule);
     const configService: ConfigService<ServerConfig, true> = app.get(ConfigService<ServerConfig, true>);
     const port: number = configService.getOrThrow<HostConfig>('HOST').PORT;
     const swagger: Omit<OpenAPIObject, 'paths'> = new DocumentBuilder()
@@ -27,6 +32,47 @@ async function bootstrap(): Promise<void> {
     });
 
     SwaggerModule.setup('docs', app, SwaggerModule.createDocument(app, swagger));
+
+    // TODO
+    if (/* frontendConfig.TRUST_PROXY*/ true !== undefined) {
+        app.set('trust proxy', 1 /*frontendConfig.TRUST_PROXY*/);
+    }
+
+    const redisConfig: RedisConfig = configService.getOrThrow<RedisConfig>('REDIS');
+    const redisClient: RedisClientType = createClient({
+        username: redisConfig.USERNAME,
+        password: redisConfig.PASSWORD,
+        socket: {
+            host: redisConfig.HOST,
+            port: redisConfig.PORT,
+            tls: redisConfig.USE_TLS,
+            key: redisConfig.PRIVATE_KEY,
+            cert: redisConfig.CERTIFICATE_AUTHORITIES,
+        },
+    });
+    await redisClient.connect();
+
+    const redisStore: RedisStore = new RedisStore({
+        client: redisClient,
+    });
+
+    app.use(
+        session({
+            store: redisStore,
+            resave: false,
+            saveUninitialized: false,
+            rolling: true,
+            cookie: {
+                maxAge: 18000, // TODO,
+                secure: false, // TODO
+            },
+            secret: 'ABCDEFG', // TODO
+        }),
+    );
+
+    app.use(passport.initialize({ userProperty: 'passportUser' }));
+    app.use(passport.session());
+    app.use(sessionAccessTokenMiddleware);
 
     await app.listen(port);
